[DesktopGL] Set GL attributes before window creation
